+++ cmd/dashgoat/config.go
@@ -34,6 +34,7 @@ type (
 		DisableDependOn       bool     `yaml:"disableDependOn"`
 		DisableMetrics        bool     `yaml:"disableMetrics"`
 		Prometheusurl         string   `yaml:"prometheusurl"`
+		HeartBeatKey          string   `yaml:"heartbeatkey"`
 		PagerdutyConfig       PdConfig `yaml:"pagerdutyconfig"`
 	}
 )
@@ -55,6 +56,9 @@ func (conf *Configer) ReadEnv() {
 	}
 	if os.Getenv("UPDATEKEY") != "" {
 		conf.UpdateKey = os.Getenv("UPDATEKEY")
+	}
+	if os.Getenv("HEARTBEATKEY") != "" {
+		conf.HeartBeatKey = os.Getenv("HEARTBEATKEY")
 	}
 	if os.Getenv("CHECKBUDDYINTERVALSEC") != "" {
 		conf.CheckBuddyIntervalSec = str2int(os.Getenv("CHECKBUDDYINTERVALSEC"))