+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [v1.6.dev] - 2024-03-19
 Add:
  - Config element logformat <txt/json>
+ - Heartbeat via HTTP GET (uri) /heartbeat/:heartbeatkey/:host/:nextupdatesec/:tags
 
 ## [v1.6.0] - 2024-03-18
 Add: